pyproject.toml
[tool.poetry]
name = "tasktracker"
version = "0.1.0"
description = ""
authors = ["Arsenii Velichkko"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
aiosqlite = "^0.19.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
uvicorn = "^0.25.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
jinja2 = "^3.1.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.13"
black = "^23.12.1"
mypy = "^1.8.0"
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240106"
alembic = "^1.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "I"]
exclude = [".git"]
line-length = 120

[tool.black]
line-length = 120

[tool.mypy]
exclude = [".git"]
ignore_missing_imports = true
plugins = "sqlalchemy.ext.mypy.plugin"